backend/requirements.txt
anyio==4.11.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
composio-core
fastapi==0.115.5
Flask==3.1.2
flask-cors==6.0.1
h11==0.14.0
h2==4.3.0
hpack==4.1.0
httpcore==0.16.3
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
numpy==2.3.4
pandas==2.3.3
pg8000==1.31.2
psutil==6.1.0
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2025.2
requests==2.32.5
rfc3986==1.5.0
scikit-learn==1.6.1
scipy==1.16.3
shap==0.49.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.31
tensorflow==2.20.0
xgboost==2.0.3
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn[standard]==0.34.0
websockets==14.1
Werkzeug==3.1.3